cdk/requirements.txt
-e .
aws-cdk.aws-ecs
aws-cdk.aws_efs
aws-cdk.aws_ec2